requirements.txt
annotated-types==0.7.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Flask==3.1.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.188.0
google-auth==2.48.0rc0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.5

